# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Carbon Credit & Stock Tracker API
  description: Comprehensive API for tracking carbon credits, stocks, and generating ML-powered forecasts
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Stock:
      type: object
      properties:
        symbol:
          type: string
          description: Stock symbol
        name:
          type: string
          description: Company name
        price:
          type: number
          description: Current price
        change:
          type: number
          description: Price change
        changePercent:
          type: number
          description: Price change percentage
        volume:
          type: integer
          description: Trading volume
        marketCap:
          type: number
          description: Market capitalization
        sector:
          type: string
          description: Industry sector
        timestamp:
          type: string
          format: date-time
...
paths:
...
  /stocks:
    get:
      summary: Get all stocks
      tags: [Stocks]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
          description: Maximum number of stocks to return
      responses:
        '200':
          description: List of stocks
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Stock'
  /stocks/{symbol}:
    get:
      summary: Get stock by symbol
      tags: [Stocks]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
          description: Stock symbol
      responses:
        '200':
          description: Stock data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Stock'
        '404':
          description: Stock not found
...